Added JWT Claims.

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.item import ItemModel
 
 
@@ -37,9 +37,13 @@ class Item(Resource):
 
         return item.json(), 201 # 201 created
 
+    @jwt_required
     def delete(self, name):
-        item = ItemModel.find_by_name(name)
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege required."}, 401
 
+        item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
             return {"message": "Item '{}' deleted.".format(name)}
